Update fixed / delete fixed
setPatch added to a .then(() => ?)
setDeleted add to a .then(() => ?)
This then allows the page to forward back to all Paddocks
The setDeleted(true) in the .then does two things:
It says that the paddock has been deleted.
And it forwards the page back to Paddocks with the deleted paddock
now gone.  Yay!

Still having trouble with PatchForm.
The old title value: title is causing the input field to freeze.