Creating a Step in the system is working
Began to add the functionality for showing all system steps

Created NewSteps (makes new steps within the subdoc of the Paddock)
Created OneStep (this will be how the steps show)
I need to create an index call for steps.
Added OneStep to the AllPaddocks file for how they will be displayed.